backend/Auth/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-environ==0.4.54
sqlparse==0.5.1
tzdata==2024.2
